reorganize BIDSFile tests
